Refs #13465 Further work on MantidMatrix Extension
