+++ Code/Mantid/MantidPlot/src/Mantid/MantidMatrixExtensionRequest.h
@@ -1,0 +1,13 @@
+#ifndef MANTIDPLOT_MANTIDMATRIXEXTENSIONREQUEST_H
+#define MANTIDPLOT_MANTIDMATRIXEXTENSIONREQUEST_H
+
+#include "MantidMatrixModel.h"
+
+class MantidMatrixExtensionRequest {
+public:
+  MantidMatrixExtensionRequest();
+  ~MantidMatrixExtensionRequest();
+  MantidMatrixTabExtension createMantidMatrixTabExtension(MantidMatrixModel::Type type);
+};
+
+#endif
